Fix x64 LuaJIT diagnostics

+++ src/lj_ccallback.c
@@ -123,7 +123,7 @@ static void *callback_mcode_init(global_State *g, uint8_t *page)
 #endif
   for (slot = 0; slot < CALLBACK_MAX_SLOT; slot++) {
     /* mov al, slot; jmp group */
-    *p++ = XI_MOVrib | RID_EAX; *p++ = (uint8_t)slot;
+    *p++ = (uint8_t)((uint8_t)XI_MOVrib | (uint8_t)RID_EAX); *p++ = (uint8_t)slot;
     if ((slot & 31) == 31 || slot == CALLBACK_MAX_SLOT-1) {
       /* push ebp/rbp; mov ah, slot>>8; mov ebp, &g. */
       *p++ = XI_PUSH + RID_EBP;
@@ -132,7 +132,7 @@ static void *callback_mcode_init(global_State *g, uint8_t *page)
       *p++ = 0x48; *p++ = XI_MOVri | RID_EBP;
       *(uint64_t *)p = (uint64_t)(g); p += 8;
 #else
-      *p++ = XI_MOVri | RID_EBP;
+      *p++ = (uint8_t)((uint8_t)XI_MOVri | (uint8_t)RID_EBP);
       *(int32_t *)p = i32ptr(g); p += 4;
 #endif
 #if LJ_64
